kata-solutions/tell-dont-ask/tell-dont-ask-python/pyproject.toml
[tool.poetry]
name = "intro-test-framework"
version = "0.1.0"
description = "Tell dont ask kata"
authors = ["Zeger Hendrikse <zegerh@yahoo.co.uk>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
pytest-watch = "^4.2.0"
pytest-cov = "4.1.0"
PyHamcrest = "^2.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

